security/HA: fix critical vulnerabilities and improve reliability
Security fixes:
- Redis TLS support, pool limits, reconnection handling
- Distributed lock ownership validation, goroutine leak fixed
- HMAC-SHA256 pub/sub signing for Signal servers
- TOCTOU fixes in login filter with UnderLock functions
- Ephemeral peer WATCH/MULTI/EXEC atomic operations
- Enhanced DSN password masking in combined mode
- CRC32 to HMAC-SHA256 for TURN credentials
- Env var rename NB_REDIS_PASSWORD to NB_HA_REDIS_PASSWORD

Reliability improvements:
- Signal peer TTL reduced for faster failover
- Timeout upper bound validation (30s cap)
- Local registry fallback when Redis lookup fails
- Signal reconnect handler with backoff and re-subscribe
- Signal liveness check and zombie cleanup

Tested: 10/10 integration tests pass, ping verified between agents.

+++ management/internals/shared/grpc/loginfilter_test.go
@@ -44,7 +44,7 @@ func (s *LoginFilterTestSuite) TestFirstLoginIsAlwaysAllowed() {
 
 	s.filter.addLogin(pubKey, meta)
 	s.Require().Contains(s.filter.logged, pubKey)
-	s.Equal(1, s.filter.logged[pubKey].sessionCounter)
+	s.Equal(uint64(1), s.filter.logged[pubKey].sessionCounter)
 }
 
 func (s *LoginFilterTestSuite) TestFlappingSameHashTriggersBan() {
